pyproject.toml
[project]
name = "goesvfi"
version = "0.4.0"
description = "GOES satellite frame interpolation tool"
dependencies = [
    "PyQt6",
    "onnxruntime",
    "numpy",
    "Pillow",
    "opencv-python-headless",
    "ffmpeg-python",
    "colorlog",
    ]
readme = "README.md"
requires-python = ">=3.11"

[project.scripts]
goesvfi = "goesvfi.cli:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-qt",
    "pytest-mock",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
python_files = "test_*.py"
addopts = "-ra -q --cov=goesvfi --cov-report=term-missing"

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88
ignore = ["E203", "W503"]
